pull logic from bucket controller to a PORO
