+++ app/controllers/employer/base_controller.rb
@@ -1,6 +1,6 @@
 class Employer::BaseController < ApplicationController
   before_action :require_employer
-
+  
   def require_employer
     unless @current_user.employer?
       render file: '/public/404', status => 404, :layout => true
