+++ app/helpers/application_helper.rb
@@ -3,10 +3,6 @@ module ApplicationHelper
     platform_admin?
   end
 
-  def new_user_path?
-    request.path == new_user_path || formatted_referrer == new_user_path
-  end
-
   def platform_admin?
     current_user && current_user.platform_admin?
   end
@@ -14,10 +10,4 @@ module ApplicationHelper
   def employer?
     current_user && current_user.employer?
   end
-
-  private
-
-    def formatted_referrer
-      '/' + request.referrer.split('/')[-2..-1].join('/') if request.referrer
-    end
 end
