+++ app/models/user.rb
@@ -3,8 +3,8 @@ class User < ApplicationRecord
   has_many :job_applications
   has_many :saved_jobs
   has_many :jobs, through: :saved_jobs
-  has_many :user_roles
   has_many :companies
+  has_many :user_roles
   has_many :roles, through: :user_roles
   validates :username, presence: true, uniqueness: true, if: "uid.nil?", on: :create
   validates :email, presence: true, uniqueness: true, if: "uid.nil?", on: :create
