+++ app/models/user_role.rb
@@ -2,3 +2,5 @@ class UserRole < ActiveRecord::Base
   belongs_to :user
   belongs_to :role
 end
+
+
