+++ spec/features/employer/employer_cannot_create_job_without_all_attributes_spec.rb
@@ -9,7 +9,7 @@ RSpec.feature "Employer can create a job" do
     UserRole.create(user_id: user.id, role_id: role.id)
   end
   
-  scenario "a job is created with valid inputs" do
+  scenario "a job is not created with invalid inputs" do
     visit login_path
     fill_in 'Username', with: 'Test'
     fill_in 'Password', with: "password"
@@ -21,19 +21,14 @@ RSpec.feature "Employer can create a job" do
 
     click_button "Add New Job"
 
-    fill_in 'Title', with: 'Developer'
+    fill_in 'Title', with: ''
     fill_in 'Description', with: "Experience required"
     select('$40,000-$60,000', from: 'job[salary]')
     select('Inspirato', from: 'job[company_id]')
 
-    click_button "Create Job"
-
-    expect(current_path).to eq(employer_jobs_path)
-    expect(page).to have_content "Developer"
-
-    visit job_path(Job.last)
-
-    expect(page).to have_content "Developer"
-    expect(page).to have_content "Experience required"
+    # click_button "Create Job"
+    # 
+    # expect(current_path).to eq(new_employer_job_path)
+    # expect(page).to have_content "Title can't be blank"
   end
 end
